Implement method GitService.updateRepository

+++ server/src/services/gitService.js
@@ -6,9 +6,13 @@ const exec = promisify(execCb);
 
 class GitService {
   async updateRepository(repoName) {
-    await fs.remove('./repo');
-    const result = await exec(`git clone https://github.com/${repoName}.git repo`);
-    return result;
+    const currentRepoName = await this.getCurrentRepositoryName();
+    if (currentRepoName === repoName) {
+      await exec('cd repo && git pull');
+    } else {
+      await fs.remove('./repo');
+      return this.cloneRepository(repoName);
+    }
   }
 
   async cloneRepository(repoName) {
@@ -16,8 +20,17 @@ class GitService {
   }
 
   async getCurrentRepositoryName() {
-    const result = await exec('cd repo && git remote get-url origin');
-    console.log('result', result);
+    try {
+      const result = await exec('cd repo && git remote get-url origin');
+      // /(?<=\/\/.+\..+\/).+\/.+(?=\.git\n?$)/ matches from url https://example.com/user-name/repo-name.git
+      // /(?<=git@.+:).+\/.+(?=\.git\n?$)/ matches from ssh git@example.com:user-name/repo-name.git
+      const matchRepoNameRegex = /(?<=\/\/.+\..+\/).+\/.+(?=\.git\n?$)|(?<=git@.+:).+\/.+(?=\.git\n?$)/;
+      const repoOriginLink = result.stdout;
+      const matchedName = repoOriginLink.match(matchRepoNameRegex);
+      return matchedName ? matchedName[0] : '';
+    } catch (error) {
+      return '';
+    }
   }
 
   async getLastCommitHash(branchName) {
